% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econioread-package.R
\docType{package}
\name{econioread-package}
\alias{econioread}
\alias{econioread-package}
\title{econioread: Read Input-Output Tables}
\description{
Provides a step-by-step interface for reading input-output tables from 'csv', 'xls', and 'xlsx' files into 'econio' input-output table objects. Handles the extraction of headers, regions, sector names, and sector types from raw spreadsheet cells, supporting both regional and multiregional input-output tables.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

Authors:
\itemize{
  \item Mizuki Uchida \email{uchidamizuki@vivaldi.net}
}

}
\keyword{internal}
