% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_read_cells}
\alias{io_table_read_cells}
\title{Import cell contents of an input-output table}
\usage{
io_table_read_cells(
  file,
  sheets = NA,
  rows_range = c(1, Inf),
  rows_exclude = integer(),
  cols_range = c(1, Inf),
  cols_exclude = integer()
)
}
\arguments{
\item{file}{Path to the input-output table file.}

\item{sheets}{A character vector of sheet names to read from. Passed to
[tidyxl::xlsx_cells()].}

\item{rows_range, cols_range}{A numeric vector of length 2 specifying the
range of rows and columns to read from.}

\item{rows_exclude, cols_exclude}{A numeric vector of row and column numbers
to exclude.}
}
\description{
`io_table_read_cells()` reads the cell contents of an input-output table.
}
