% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_read_data}
\alias{io_table_read_data}
\title{Create data of an input-output table}
\usage{
io_table_read_data(cells, scale, check_total = TRUE)
}
\arguments{
\item{cells}{A data frame containing the cell contents of an input-output
table.}

\item{scale}{A scalar numeric specifying the scale for the numeric values.}

\item{check_total}{A scalar logical indicating whether to check the total of
the input-output table. Passed to [econio::io_table_multiregional()] or
[econio::io_table_regional()].}
}
\description{
`io_table_read_data()` creates data of an input-output table.
}
