% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_read_regions}
\alias{io_table_read_regions}
\title{Create columns of regions in an input-output table}
\usage{
io_table_read_regions(cells, input_region_glue, output_region_glue)
}
\arguments{
\item{cells}{A data frame containing the cell contents of an input-output
table.}

\item{input_region_glue, output_region_glue}{A scalar character string
specifying the glue pattern for input and output regions.}
}
\description{
`io_table_read_regions()` creates columns of regions in an input-output
table.
}
