% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_read_sector_names}
\alias{io_table_read_sector_names}
\title{Create columns of sector names in an input-output table}
\usage{
io_table_read_sector_names(
  cells,
  input_sector_name_glue,
  output_sector_name_glue
)
}
\arguments{
\item{cells}{A data frame containing the cell contents of an input-output
table.}

\item{input_sector_name_glue, output_sector_name_glue}{A scalar character
speecifying the glue pattern for input and output sector names.}
}
\value{
A data frame containing the sector names of an input-output table.
}
\description{
`io_table_read_sector_names()` creates columns of sector names in an
input-output table.
}
