% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_reader}
\alias{io_table_reader}
\title{Input-output table reader}
\usage{
io_table_reader(file, region_type = c("regional", "multiregional"))
}
\arguments{
\item{file}{Path to the input-output table file.}

\item{region_type}{Type of the input-output table. Either `"regional"` or
`"multiregional"`, by default `"regional"`.}
}
\description{
`io_table_reader()` provides a step-by-step process to read an input-output
table from a file.
}
