% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverbial.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_step}
\alias{end_step}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{adverbial}{\code{\link[adverbial]{as_step}}, \code{\link[adverbial]{end_step}}}
}}

