% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econioread-package.R
\docType{package}
\name{econioread-package}
\alias{econioread}
\alias{econioread-package}
\title{econioread: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
