% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econioread-package.R
\docType{package}
\name{econioread-package}
\alias{econioread}
\alias{econioread-package}
\title{econioread: Read Input-Output Tables}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
