% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-read.R
\name{io_table_read_data}
\alias{io_table_read_data}
\title{Create data of an input-output table}
\usage{
io_table_read_data(cells, scale, total_tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{cells}{A data frame containing the cell contents of an input-output
table.}

\item{scale}{A scalar numeric specifying the scale for the numeric values.}

\item{total_tolerance}{Passed to [econio::io_table_multiregional()] or
[econio::io_table_regional()]. By default, `.Machine$double.eps^0.5`.}
}
\value{
An input-output table object.
}
\description{
`io_table_read_data()` creates data of an input-output table.
}
